Prepare to refactor decompression
