+++ components/core/src/streaming_archive/reader/CombinedLogtypeTable.cpp
