+++ components/core/src/streaming_archive/reader/LogtypeMetadata.cpp
