+++ components/core/src/streaming_archive/reader/LogtypeTable.cpp
