+++ components/core/src/streaming_archive/reader/LogtypeTableManager.cpp
